Move lib/config into shared/config instead, as it really is a shared file
