+++ webapplication/lib/errors.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var config = require('./config');
+var config = require('../../shared/config');
 
 /**
  * Application errors
