+++ webapplication/lib/pug-resolve-plugin.js
@@ -4,7 +4,7 @@ var pugLoad = require('pug-load');
 
 module.exports = function(config) {
   if(!config) {
-    config = require('./config');
+    config = require('../../shared/config');
   }
 
   var viewsPaths = config.appPaths.map((p) => {
