+++ webapplication/lib/routes.js
@@ -9,7 +9,7 @@ const motifTagging = require('./controllers/motif-tagging');
 const index = require('./controllers/index');
 const doc = require('./controllers/document');
 const series = require('./controllers/series');
-const config = require('./config');
+const config = require('../../shared/config');
 const indexing = require('../controllers/indexing');
 const users = require('../controllers/users');
 const keystonePlugin = require('../keystone');
