+++ webapplication/start.js
@@ -1,5 +1,5 @@
 // TODO: Config must have customization set as very first because some modules depend on config being complete at require time (bad, shouldfix)
-const config = require('./lib/config');
+const config = require('../shared/config');
 config.setCustomizationPath(__dirname + '/..');
 
 const co = require('./server');
